Fix window size not shrinking
Another round of fighting against GTK's automatic size handling.

get_size() doesn't reflect the latest window size but just the current.
As a result, inside the move_resize() method, the calculated image scale
will be consequently wrong.

Not not use get_size(), but keep track of the target window size
manually. Update these on actual window size changes.

Disable the entire window's height-for-width management instead of doing
that on just the label. This ensures the window never negotiates the
size with any nested widget (which includes the image!)